+++ lib/semian/circuit_breaker.rb
@@ -6,16 +6,24 @@ module Semian
 
     def_delegators :@state, :closed?, :open?, :half_open?
 
-    attr_reader :name, :half_open_resource_timeout, :error_timeout, :state, :last_error
+    attr_reader(
+      :name,
+      :half_open_resource_timeout,
+      :error_timeout,
+      :state,
+      :last_error,
+      :error_threshold_timeout_enabled,
+    )
 
     def initialize(name, exceptions:, success_threshold:, error_threshold:,
       error_timeout:, implementation:, half_open_resource_timeout: nil,
-      error_threshold_timeout: nil)
+      error_threshold_timeout: nil, error_threshold_timeout_enabled: true)
 
       @name = name.to_sym
       @success_count_threshold = success_threshold
       @error_count_threshold = error_threshold
       @error_threshold_timeout = error_threshold_timeout || error_timeout
+      @error_threshold_timeout_enabled = error_threshold_timeout_enabled.nil? ? true : error_threshold_timeout_enabled
       @error_timeout = error_timeout
       @exceptions = exceptions
       @half_open_resource_timeout = half_open_resource_timeout
@@ -131,7 +139,10 @@ module Semian
     end
 
     def push_time(window, time: Time.now)
-      window.reject! { |err_time| err_time + @error_threshold_timeout < time.to_i }
+      if error_threshold_timeout_enabled
+        window.reject! { |err_time| err_time + @error_threshold_timeout < time.to_i }
+      end
+
       window << time.to_i
     end