+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Typo in error message for missing option `:tickets`. (#412)
 * Add process id in DEBUG message for bulkheading/resource. (#416)
+* Feature: Add ability to turn off window based circuit eviction (#418)
 
 # v0.15.0